requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & API Clients
httpx==0.25.1
requests==2.31.0
websockets==12.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning / Stats
scikit-learn==1.3.2
statsmodels==0.14.0

# News & NLP
tweepy==4.14.0
newsapi-python==0.2.7
nltk==3.8.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Weather Data
noaa-sdk==0.1.18

# Telegram
python-telegram-bot==20.7

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
python-json-logger==2.0.7

# Time & Date
python-dateutil==2.8.2
pytz==2023.3
arrow==1.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-mock==0.7.0
freezegun==1.4.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.5.0

# Utilities
tenacity==8.2.3  # Retry logic
cachetools==5.3.2  # Caching
joblib==1.3.2  # Parallel processing
tqdm==4.66.1  # Progress bars
click==8.1.7  # CLI tools
